requirements.txt
langchain
langchain-community
google-generativeai
langchain-google-genai
pymongo
flask
scikit-learn
numpy
pdf2image
opencv-python
pytesseract
python-magic
redis
celery
flask-caching
flask-mail